Make the bot not stride to far away from start pos.
After a little while the bot will now select a fort close to starting
position as the next fort. This makes sure the bot stays in close
proximity to the starting location. To make sure the bot does not
allways use the same stops it will now select them randomly.